requirements.txt
# Core dependencies
Flask==3.0.0
PyYAML==6.0.1
markdown==3.5.1
requests==2.31.0
feedparser==6.0.11

# Static site generation for GitHub Pages
Frozen-Flask==1.0.1

# Development dependencies
pytest==7.4.3
pytest-flask==1.3.0
black==23.12.1
flake8==7.0.0
